requirements.txt
hydra-core==1.2.0
tqdm==4.64.0
bs4==0.0.1
tldextract==3.3.1
warcio==1.7.4
networkx==2.8.6
pyvis==0.3.0
coloraide==1.4
fasttext==0.9.2
fastapi==0.85.0
pyyaml==5.4.1
pyserini
faiss-cpu==1.7.2
pysbd==0.3.4
requests==2.27.1
pandas
omegaconf
numpy
scikit-learn
matplotlib
